+++ src/ObjectServer.Core/Model/Fields/IFieldCollection.cs
@@ -9,7 +9,7 @@ namespace ObjectServer.Model
     {
         IField Integer(string name);
         IField BigInteger(string name);
-        IField Float(string name);
+        IField Double(string name);
         IField Decimal(string name);
         IField Boolean(string name);
         IField DateTime(string name);
