+++ src/ObjectServer.DemoModule/Demo/EmployeeModel.cs
@@ -15,7 +15,7 @@ namespace ObjectServer.DemoModule
         {
             Fields.Chars("name").SetLabel("姓名").Required();
             Fields.Chars("address").SetLabel("地址");
-            Fields.Decimal("salary").SetLabel("月薪");
+            Fields.Double("salary").SetLabel("月薪");
             Fields.Date("birthdate").SetLabel("出生日期");
         }
     }
